+++ infrastructure/src/main/kotlin/KotlinLibraryPlugin.kt
@@ -1,32 +1,37 @@
 package com.redmadrobot.build
 
-import com.redmadrobot.build.extension.redmadrobotExtension
+import com.redmadrobot.build.extension.TestOptions
+import com.redmadrobot.build.internal.*
 import org.gradle.api.JavaVersion
-import org.gradle.api.Plugin
 import org.gradle.api.Project
+import org.gradle.api.tasks.testing.Test
 import org.gradle.kotlin.dsl.apply
 import org.gradle.kotlin.dsl.repositories
+import org.gradle.kotlin.dsl.withType
 
-public class KotlinLibraryPlugin : Plugin<Project> {
-    override fun apply(target: Project) {
-        with(target) {
-            requireRootPlugin()
-            apply(plugin = "kotlin")
+public class KotlinLibraryPlugin : InfrastructurePlugin() {
 
-            java {
-                targetCompatibility = JavaVersion.VERSION_1_8
-                sourceCompatibility = JavaVersion.VERSION_1_8
-            }
+    override fun Project.configure() {
+        apply(plugin = "kotlin")
 
-            kotlin.explicitApi()
-
-            configureKotlin()
-            configureKotlinTest()
-            configureRepositories()
-            configureKotlinDependencies()
-            configureKotlinTestDependencies(redmadrobotExtension.test)
+        java {
+            targetCompatibility = JavaVersion.VERSION_1_8
+            sourceCompatibility = JavaVersion.VERSION_1_8
         }
+
+        kotlin.explicitApi()
+
+        val extension = redmadrobotExtension
+        configureKotlin()
+        configureKotlinDependencies(extension.kotlinVersion)
+        configureKotlinTest(extension.test)
+        configureKotlinTestDependencies(extension.kotlinVersion, extension.test)
+        configureRepositories()
     }
+}
+
+private fun Project.configureKotlinTest(options: TestOptions) {
+    tasks.withType<Test>().configureEach { setTestOptions(options) }
 }
 
 private fun Project.configureRepositories() {