+++ infrastructure/src/main/kotlin/PublishPlugin.kt
@@ -1,77 +1,68 @@
 package com.redmadrobot.build
 
 import com.redmadrobot.build.extension.isReleaseVersion
-import com.redmadrobot.build.extension.redmadrobotExtension
-import org.gradle.api.Plugin
+import com.redmadrobot.build.internal.java
 import org.gradle.api.Project
+import org.gradle.api.publish.PublishingExtension
 import org.gradle.api.publish.maven.MavenPublication
-import org.gradle.jvm.tasks.Jar
 import org.gradle.kotlin.dsl.*
 import org.gradle.plugins.signing.Sign
+import org.gradle.plugins.signing.SigningExtension
 
-public class PublishPlugin : Plugin<Project> {
+public open class PublishPlugin : InfrastructurePlugin() {
 
     public companion object {
         public const val PUBLICATION_NAME: String = "maven"
         public const val PLUGIN_PUBLICATION_NAME: String = "pluginMaven"
     }
 
-    override fun apply(target: Project) {
-        with(target) {
-            apply(plugin = "maven-publish")
+    protected val publishing: PublishingExtension
+        get() = project.extensions.getByName<PublishingExtension>("publishing")
 
-            val publicationName = when {
-                plugins.hasPlugin("kotlin-android") -> configureAndroidPublication()
-                plugins.hasPlugin("java-gradle-plugin") -> configurePluginPublication()
-                else -> configurePublication()
+    protected fun publishing(configure: PublishingExtension.() -> Unit) {
+        project.extensions.configure("publishing", configure)
+    }
+
+    private fun signing(configure: SigningExtension.() -> Unit) {
+        project.extensions.configure("signing", configure)
+    }
+
+    override fun Project.configure() {
+        apply(plugin = "maven-publish")
+
+        val publicationName = when {
+            plugins.hasPlugin("kotlin-android") -> configureAndroidPublication()
+            plugins.hasPlugin("java-gradle-plugin") -> configurePluginPublication()
+            else -> configurePublication()
+        }
+
+        // Do it after project evaluate to be able to access publications created later
+        afterEvaluate {
+            val options = redmadrobotExtension.publishing
+
+            publishing.publications.getByName<MavenPublication>(publicationName) {
+                pom {
+                    name.convention(project.name)
+                    description.convention(project.description)
+                    options.configurePom(this)
+                }
             }
 
-            // Do it after project evaluate to be able to access publications created later
-            afterEvaluate {
-                val redmadrobot = redmadrobotExtension
-
-                publishing.publications.getByName<MavenPublication>(publicationName) {
-                    pom {
-                        name.convention(project.name)
-                        description.convention(project.description)
-                        redmadrobot.publishing.configurePom(this)
-                    }
-                }
-
-                if (redmadrobot.publishing.signArtifacts) {
-                    configureSigning(publicationName, redmadrobot.publishing.useGpgAgent)
-                }
+            if (options.signArtifacts) {
+                configureSigning(publicationName, options.useGpgAgent)
             }
         }
     }
 
-    private fun Project.configureAndroidPublication(): String {
-        val sourcesJar = tasks.register<Jar>("sourcesJar") {
-            archiveClassifier.set("sources")
-            from(android.sourceSets["main"].java.srcDirs)
-        }
-
-        // Pre-create publication to make it configurable
-        publishing.publications.create<MavenPublication>(PUBLICATION_NAME)
-
-        // Because the components are created only during the afterEvaluate phase, you must
-        // configure your publications using the afterEvaluate() lifecycle method.
-        afterEvaluate {
-            if (version == Project.DEFAULT_VERSION) {
-                version = checkNotNull(android.defaultConfig.versionName) {
-                    "You should specify either project 'version' or 'android.versionName' for publication."
-                }
-            }
-
-            publishing {
-                publications.getByName<MavenPublication>(PUBLICATION_NAME) {
-                    from(components["release"])
-                    artifact(sourcesJar.get())
-                }
-            }
-        }
-
-        return PUBLICATION_NAME
+    protected open fun Project.configureAndroidPublication(): String {
+        error(
+            """
+            The project was recognized as Android-related because the plugin 'kotlin-android' was found.
+            If you want to publish android artifacts, you should use 'infrastructure-android'
+            instead of 'infrastructure'.
+            You can change it in gradle.settings.kts
+            """.trimIndent(),
+        )
     }
 
     private fun Project.configurePluginPublication(): String {