+++ infrastructure/src/main/kotlin/internal/Configurations.kt
@@ -1,0 +1,7 @@
+@file:Suppress("TopLevelPropertyNaming")
+
+package com.redmadrobot.build.internal
+
+internal const val api: String = "api"
+internal const val testImplementation: String = "testImplementation"
+internal const val detektPlugins: String = "detektPlugins"